# src/internal/docs/swagger.yaml
basePath: /
definitions:
...
  document_handler.RequestID:
    properties:
      ID:
        type: string
    type: object
  document_handler.RequestPassed:
    properties:
      ID:
        type: string
      has_passed:
        type: boolean
    type: object
...
  models.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Sender: Role check depends on the order
    x-enum-varnames:
    - Sender
    - Controller
    - Admin
...
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  user_handler.RequestChangeRole:
    properties:
      login:
        type: string
      req_role:
        $ref: '#/definitions/models.Role'
    type: object
...
host: localhost:8080
info:
  contact:
    email: ggwpezsmth
    name: Andrew
    url: http://www.swagger.io/support
  description: An App for storing, getting annots and auto checking documents.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Annoter
  version: "1.0"
paths:
...
  /document/getDocument:
    post:
      consumes:
      - application/json
      description: Fetches a document file without metadata by its ID
      parameters:
      - description: Document ID
        in: body
        name: Document
        required: true
        schema:
          $ref: '#/definitions/document_handler.RequestID'
      produces:
      - application/pdf
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get document by ID
      tags:
      - Document
  /document/getDocumentsMeta:
    get:
      consumes:
      - application/json
      description: Gets all document metadata which was created by this user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get document meta data
      tags:
      - Document
  /document/getReport:
    post:
      consumes:
      - application/json
      description: Fetches a report file without metadata by its ID
      parameters:
      - description: Report ID
        in: body
        name: ReportID
        required: true
        schema:
          $ref: '#/definitions/document_handler.RequestID'
      produces:
      - application/pdf
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get report by ID
      tags:
      - Document
  /document/makeDecision:
    post:
      consumes:
      - application/json
      description: Set's a field of the document has passed to true
      parameters:
      - description: id and bool value of whether the lab was merged
        in: body
        name: HasPassedParams
        required: true
        schema:
          $ref: '#/definitions/document_handler.RequestPassed'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Mark a document to pass by normocontroller (requires a role to be a
        normocontroller)
      tags:
      - Document
  /document/report:
    post:
      consumes:
      - multipart/form-data
      description: Gets a document, saves it with metadata on the system, then creates
        a report,
      parameters:
      - description: Document file to report
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/pdf
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create an error report by given document
      tags:
      - Document
...
  /user/getUsers:
    get:
      consumes:
      - application/json
      description: Get all users parametersm available if you are admin
      produces:
      - application/json
      responses:
        "200":
          description: some internal user errror
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all users data (available if admin)
      tags:
      - Users
  /user/role:
    post:
      consumes:
      - application/json
      description: Set user role, only available if you are admin
      parameters:
      - description: data to get the user and change his role
        in: body
        name: ChangeUserRoleParams
        required: true
        schema:
          $ref: '#/definitions/user_handler.RequestChangeRole'
      produces:
      - application/json
      responses:
        "200":
          description: Annotation not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Change user role (available if admin)
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"